wip readme

+++ src/AdapterLocal.ts
@@ -270,6 +270,9 @@ export class AdapterLocal extends AbstractAdapter {
 
     try {
       const p = path.join(this._config.directory, bucketName, fileName);
+      if (!fs.existsSync(p)) {
+        return { value: "ok", error: null };
+      }
       await fs.promises.unlink(p);
       return { value: "ok", error: null };
     } catch (e) {
